refactor: 가격 double -> int 변경

+++ backend/src/main/java/com/caffe/domain/purchase/dto/req/PurchaseReqDto.java
@@ -9,12 +9,12 @@ public record PurchaseReqDto(
         int productId,
 
         @Positive
-        double price,
+        int price,
 
         @Positive
         int quantity,
 
         @Positive
-        double totalPrice
+        int totalPrice
 ) {
 }
